Support FilteredAdapter interface.
